Alpha Submission
at this time the table view is able to populate all workouts added the the list. both UI elements for the apple watch and the iphone has been formatted correctly.

whats next?
I must create a way to update the data for water intake then connect the data from the iphone and apple watch for one value to show